Refactor(Employee): Ganti field primitif tanggal masuk dengan LocalDate
Mengatasi bad smell Primitive Obsession dengan mengganti field integer 'yearJoined', 'monthJoined', dan 'dayJoined' dengan satu field 'LocalDate joinDate' di kelas Employee.
