--- TaxCalculator/src/module-info.java
@@ -1,6 +1,0 @@
-module TaxCalculator {
-	
-	
-	
-	
-}
