+++ bin/module-info.class
