+++ src/module-info.java
@@ -1,0 +1,6 @@
+module TaxCalculator {
+	
+	
+	
+	
+}
